db file
This code connects to a MySQL database and inserts data into a table called "translations". The data includes the language, photo path, original text, and translated text. The language, photo path, original text, and translated text are obtained using functions from another module called "ImageOCR_trans". Once the data is obtained, it is inserted into the table using SQL. Finally, the changes are committed to the database and a message is printed to indicate the success of the operation.